pyproject.toml
[tool.poetry]
name = "sim-lang"
version = "0.1.0"
description = "benjamin asset simulation langugage"
authors = ["Kyoodong <wind.dong.dream@gmail.com\">"]

[tool.poetry.dependencies]
python = "^3.8"
PyMySQL = "^1.0.2"
pytest = "^6.2.2"
python-dotenv = "^0.15.0"
pandas = "^1.2.1"
numpy = "^1.20.1"
utils = { git = "https://github.com/benjamin-asset/utils.git" }
boto3 = "^1.17.27"
anytree = "^2.8.0"


[tool.poetry.dev-dependencies]
pytest = "^6.2.2"

[tool.black]
line-length = 119
target-version = ['py38']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
